git-svn-id: file:///srv/local/data/cleanRepo/trunk@6987 a0517364-e873-4167-b1c8-416daa5a6b91

+++ matsim/src/playground/johannes/graph/mcmc/GibbsSampler.java
@@ -52,11 +52,11 @@ public class GibbsSampler {
 		this.inteval = interval;
 	}
 	
-	public void sample(AdjacencyMatrix y, ConditionalDistribution d, int burninTime, MCMCSampleDelegate handler) {
+	public void sample(AdjacencyMatrix y, ConditionalDistribution d, long burninTime, MCMCSampleDelegate handler) {
 		long time = System.currentTimeMillis();
 		
 		int accept = 0;
-		for(int it = 0; it < burninTime; it++) {
+		for(long it = 0; it < burninTime; it++) {
 			if(step(y, d))
 				accept++;
